Extracting DependsOnMethod to a dedicated file.
