+++ src/routing.ts
@@ -1,8 +1,9 @@
 import { Express, static as _serveStatic } from "express";
 import { Logger } from "winston";
 import { CommonConfig } from "./config-type";
-import { AbstractEndpoint, Endpoint } from "./endpoint";
-import { DependsOnMethodError, RoutingError } from "./errors";
+import { DependsOnMethod } from "./depends-on-method";
+import { AbstractEndpoint } from "./endpoint";
+import { RoutingError } from "./errors";
 import { AuxMethod, Method } from "./method";
 import { getStartupLogo } from "./startup-logo";
 
@@ -12,40 +13,6 @@ export type StaticHandler = ReturnType<typeof _serveStatic> & {
 
 export const serveStatic = (...params: Parameters<typeof _serveStatic>) =>
   _serveStatic(...params) as StaticHandler;
-
-export class DependsOnMethod {
-  constructor(
-    public readonly methods: {
-      [K in Method]?:
-        | Endpoint<any, any, any, any, K, any, any>
-        | Endpoint<any, any, any, any, Method, any, any>;
-    }
-  ) {
-    (Object.keys(methods) as (keyof typeof methods)[]).forEach((key) => {
-      if (key in methods) {
-        const endpointMethods = methods[key]?.getMethods() || [];
-        if (!endpointMethods.includes(key)) {
-          throw new DependsOnMethodError(
-            `The endpoint assigned to the '${key}' parameter must have at least this method in its specification.\n` +
-              "This error should prevent mistakes during the development process.\n" +
-              "Example:\n\n" +
-              `new ${this.constructor.name}({\n` +
-              `  ${key}: endpointsFactory.build({\n` +
-              `    methods: ['${key}', ` +
-              ((methods[key]?.getMethods() || [])
-                .map((m) => `'${m}'`)
-                .join(", ") || "...") +
-              "]\n" +
-              `    // or method: '${key}'\n` +
-              "    ...\n" +
-              "  })\n" +
-              "});\n"
-          );
-        }
-      }
-    });
-  }
-}
 
 export interface Routing {
   [SEGMENT: string]:
